Risk Quantification Project
In this project, I estimate the Value-at-Risk and the Expected Shortfall or Tail Value-at-Risk through the variance-covariance method, the historical simulation method, and the Monte Carlo method. In the latter, I use different distributions applied in the field of risk management and I also use of different types of copulae. Finally, I carry on a factorial analysis by estimating a factorial model, and I do a brief extreme value distribution analysis to estimate different risk measures.