Volatility Modeling project
In this project, I analyze the time series of prices and returns of the Nikkei225 index in order to model both the mean and the volatility. Then, diverse GARCH models are used for forecasting, estimating volatility for the selected time horizon, estimating a news impact curve, and forecasting VaR for out-of-sample observations. I also compare the estimated volatility from the GARCH models with the one obtained from EWMA and the historical volatility estimator, and I use the FTSE100 series to estimate a bivariate DCC-GARCH model and analyze correlations and news impacts on correlations.